==> step4_cosine_addfeatures.ipynb <==
import numpy as np
import pandas as pd
import nltk
import os
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text  import CountVectorizer
from sklearn.metrics.pairwise import linear_kernel
from sklearn.metrics.pairwise import cosine_similarity

os.chdir('/Users/knar/Desktop/ComputationalLabs/Project/Data')

#pd.set_option('display.max_colwidth', -1)
#pd.set_option('display.max_colwidth', 10)

##  Combine 3 datasets - does not need to be run again

data = pd.read_csv("step3processed.csv")

rawdata = data[:]

##  Cosine Similarity

#Can set parameters for how Tfidf is calculated, update to have a token as 1+ alphanumeric factor and not >1
#this way single character words (like I, F for female etc. are included)

def getCos(df,passed_dict):
    notelist = list(df.TEXT)
    rowidlist = list((df.ROW_ID))
    if len(notelist) <=1:
        passed_dict[rowidlist[0]] = "Only 1 note"
        return(passed_dict)
    tfidf = TfidfVectorizer(analyzer ="word", token_pattern='(?u)\\b\\w+\\b')
    tfs = tfidf.fit_transform(notelist)
    feature_names = tfidf.get_feature_names()
    
#loop through each note, calculate cosine similarity against it against all others (including self)
#to get average, make sure to subtact comparison to itself (1) in numerator (cosine) and denominator (number of notes)
#since first using tfidf (produces normalized vectors) can  use linear_kernal (faster) instead of cosine_similarity
# did test below as well to make sure same results
    counter = 0
    for note in tfs:
        row_id = rowidlist[counter]
        cosine_similarities = linear_kernel(note, tfs).flatten()
        avgcos = (cosine_similarities.sum()-1)/(len(cosine_similarities)-1)
        passed_dict[row_id] = avgcos
        counter +=1
    return(passed_dict)

#loop through each patient one at a time
#loop through patient's category types
# caluclate cosine similarity for patient/category subset of notes and add to categ dict by the rowid
#then loop through all the descriptions within category and add to desc dict by rowid
cosine_desc_dict = {}
cosine_categ_dict = {}
count = 0

for subj in data.SUBJECT_ID.unique():
    if count%100000 == 0:
            print(count)
    pat_df = data[data.SUBJECT_ID == subj]
    for cat in pat_df.CATEGORY.unique():
        cat_df = pat_df[pat_df.CATEGORY == cat]
        getCos(cat_df,cosine_categ_dict)
            
        for desc in cat_df.DESCRIPTION.unique():
            desc_df = cat_df[cat_df.DESCRIPTION == desc]
            getCos(desc_df,cosine_desc_dict)
            count+=1


#append the description cosine similarity for each row
data["CosDesc"] = data.ROW_ID.map(cosine_desc_dict)
#[cosine_desc_dict[x] for x in data.ROW_ID]

#append the category cosine similarity for each row
data["CosCat"] = data.ROW_ID.map(cosine_categ_dict)
#[cosine_categ_dict[x] for x in data.ROW_ID]

#data.to_csv("/Users/klarow/Desktop/ComputationalLabs/Project/Data/cosine_clean.csv",index = False)

## Test Train and Merge with HADM-ID

addmissions_data = pd.read_csv("ADMISSIONS.csv")

addmissions_data.ADMISSION_LOCATION.unique()

#subset on columns we want to include in analysis (ie admission type, adm location, insurannce, lang, ethnicity)
to_add_hadm = addmissions_data[[2,6,7,8,9,10,11,13]]

to_add_hadm.head()

#Assume that missing values are english
to_add_hadm.LANGUAGE.fillna(value="ENGL",inplace=True)

#number of instances with no HADMID
len(data[data.HADM_ID.isnull()==True])

#remove these instances with no hadmid, since so many variables are dependent on matching on it
data = data[data.HADM_ID.isnull()==False]

dfwith_admin = pd.merge(data,to_add_hadm,on = "HADM_ID")

data = dfwith_admin

##  Partition Dates by Hours

#get times from date time
ctime1 = data.CHARTTIME.str.split(" ",expand = True)

#breakup hour, minutes, seconds (in military time)
ctime2 = ctime1[1].str.split(":",expand = True)

#take the hour as the category
ctime3 = ctime2[0]

#create new column with the chart time as category (24 hrs)
data["CHARTTIMECAT"] = ctime3

#repeat steps for store time
stime = data.STORETIME.str.split(" ",expand = True)

stime2 = stime[1].str.split(":", expand = True)

#stime2[0].unique()

stime3 = stime2[0]

data["STORETIMECAT"] = ctime3

data.head()

##  Partition  Test/Train  and set all blank iserror to 0

data.ISERROR.fillna(value = 0,inplace=True)

errortemp = data[data.ISERROR == 1]
errortemp["Train"] = np.random.normal(0,1,len(errortemp),) <= .80
#errortemp.Train
correcttemp = data[data.ISERROR != 1]
correcttemp["Train"] = np.random.normal(0,1,len(correcttemp),) <= .80
#correcttemp.Train
#mr (model ready) dataframe
data = pd.concat([correcttemp,errortemp], axis = 0)

##  Reduce # Language Categories, combine all those with < 100 notes with subject speaking it

data["OldLanguage"] = data["LANGUAGE"]

#data.head()

#get counts of # people who speak each language
language_Df = pd.DataFrame(data.LANGUAGE.value_counts()).reset_index()

#create dictionary to collapse languages with fewer than 100 people speaking it into an other category
lang_dict = {}
for each in language_Df[language_Df.LANGUAGE < 100]["index"]:
    lang_dict[each] = "Other"

#keep other languages as is
for each in language_Df[language_Df.LANGUAGE >= 100]["index"]:
    lang_dict[each] = each

#assign other to those languages with fewer
data["LANGUAGE"] = data.LANGUAGE.map(lang_dict)


#columns to drop, not needed for modelling
to_drp = ["CHARTDATE","CHARTTIME","STORETIME","TIMECHARTED","TIMESTORED"]

#drop
data_clean = data.drop(to_drp, axis=1, level=None, inplace=False, errors='raise')

#output modelread file
data_clean.to_csv("/Users/klarow/Desktop/ComputationalLabs/Project/Data/modelready.csv",index = False)

## Test Code

# cosine_similarities = linear_kernel(tfs[1:2], tfs).flatten()
# avgcos = (cosine_similarities.sum()-1)/(len(cosine_similarities)-1)
# cosine_similarities
# avgcos
